+++ src/imio/dataexchange/db/__init__.py
@@ -9,3 +9,8 @@ from zope.sqlalchemy import ZopeTransactionExtension
 
 DBSession = scoped_session(sessionmaker(extension=ZopeTransactionExtension()))
 DeclarativeBase = declarative_base()
+
+
+def temporary_session(engine):
+    """Return a session that can be used without transaction manager"""
+    return scoped_session(sessionmaker(bind=engine))